updates v1.3

+++ courses/models.py
@@ -7,8 +7,8 @@ from teachers.models import Teacher
 class Course(models.Model):
     title = models.CharField(max_length=100)
     description = models.TextField()
-    duration = models.IntegerField(null=True)
-    credit = models.IntegerField(null=True)
+    duration = models.DecimalField(max_digits=10, decimal_places=2, null=True)
+    credit = models.DecimalField(max_digits=10, decimal_places=2, null=True)
     department = models.CharField(max_length=100, null=True)
     image = models.ImageField(upload_to='courses/images/', blank=True)
     teacher = models.ForeignKey(Teacher, on_delete=models.CASCADE)
